Rename firebase-resource-manager.ts to projectsCreate.ts
